# openapi.yaml
openapi: 3.0.3
info:
  title: Claims Management System API
  description: |-
    API specification for the Claims Management System, a Spring Boot application for managing insurance claims.
    
    ## Features
    - Create and manage claimants
    - Submit and track insurance claims
    - Assign claims to staff members
    - Monitor workload and claims status
    
    ## Claim Lifecycle States
    - SUBMITTED: Initial claim submission
    - ASSIGNED: Claim assigned to a staff member
    - ASSESSMENT_IN_PROGRESS: Under review
    - MORE_INFO_REQUESTED: Additional information needed
    - APPROVED: Claim approved
    - REJECTED: Claim rejected
    - SETTLEMENT_IN_PROGRESS: Settlement being processed
    - CLOSED: Claim closed
  contact:
    name: API Support
    email: support@chubb.com
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.claimsmanagement.com
    description: Production server
...
paths:
...
  /api/v1/claims:
    post:
      tags:
        - Claims
      summary: Create a new claim
      description: Submit a new insurance claim for a claimant
      operationId: createClaim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClaimRequest'
            examples:
              motor:
                value:
                  claimantMemberNumber: "CM-1001"
                  claimType: MOTOR
                  description: Car accident on Main Street involving two vehicles
              property:
                value:
                  claimantMemberNumber: "CM-1002"
                  claimType: PROPERTY
                  description: Water damage to kitchen from burst pipe
      responses:
        '201':
          description: Claim created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseClaimResponse'
              examples:
                default:
                  value:
                    success: true
                    data:
                      claimNumber: "CLM-ABC12345"
                      claimantMemberNumber: "CM-1001"
                      claimType: MOTOR
                      status: SUBMITTED
                      description: Car accident on Main Street involving two vehicles
                      createdAt: "2024-01-15T10:30:00"
                      updatedAt: "2024-01-15T10:30:00"
                    message: Claim created successfully
                    timestamp: "2024-01-15T10:30:00.000Z"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: Claimant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'

    get:
      tags:
        - Claims
      summary: Get claims by claimant
      description: Retrieve all claims for a specific claimant by internal ID or member number
      operationId: getClaimsByClaimant
      parameters:
        - name: claimantId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Internal claimant ID to filter claims
          example: 11111111-1111-1111-1111-111111111111
        - name: claimantMemberNumber
          in: query
          required: false
          schema:
            type: string
          description: Claimant member number to filter claims
          example: CM-1001
      responses:
        '200':
          description: List of claims retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseClaimResponseList'
              examples:
                default:
                  value:
                    success: true
                    data:
                      - claimNumber: "CLM-ABC12345"
                        claimantMemberNumber: "CM-1001"
                        claimType: MOTOR
                        status: SUBMITTED
                        description: Car accident on Main Street
                        createdAt: "2024-01-15T10:30:00"
                        updatedAt: "2024-01-15T10:30:00"
                      - claimNumber: "CLM-DEF67890"
                        claimantMemberNumber: "CM-1001"
                        claimType: PROPERTY
                        status: ASSIGNED
                        description: Water damage to kitchen
                        createdAt: "2024-01-14T15:45:00"
                        updatedAt: "2024-01-15T09:00:00"
                    message: OK
                    timestamp: "2024-01-15T10:30:00.000Z"
        '400':
          description: Missing required parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/claims/{claimNumber}:
    get:
      tags:
        - Claims
      summary: Get claim by claim number
      description: Retrieve a specific claim's details using the generated claim number
      operationId: getClaim
      parameters:
        - name: claimNumber
          in: path
          required: true
          schema:
            type: string
          description: Customer-facing claim number
          example: CLM-ABC12345
      responses:
        '200':
          description: Claim found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseClaimResponse'
              examples:
                default:
                  value:
                    success: true
                    data:
                      claimNumber: "CLM-ABC12345"
                      claimantMemberNumber: "CM-1001"
                      claimType: MOTOR
                      status: SUBMITTED
                      description: Car accident on Main Street involving two vehicles
                      createdAt: "2024-01-15T10:30:00"
                      updatedAt: "2024-01-15T10:30:00"
                    message: OK
                    timestamp: "2024-01-15T10:30:00.000Z"
        '404':
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/claims/{claimNumber}/assign:
    post:
      tags:
        - Claims
      summary: Assign claim to staff member
      description: Assign a claim to a staff member for processing
      operationId: assignClaim
      parameters:
        - name: claimNumber
          in: path
          required: true
          schema:
            type: string
          description: Public claim number to assign
          example: CLM-ABC12345
        - name: staffNumber
          in: query
          required: true
          schema:
            type: string
          description: Public staff number to assign the claim to
          example: STAFF-1001
      responses:
        '200':
          description: Claim assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseClaimResponse'
              examples:
                default:
                  value:
                    success: true
                    data:
                      claimNumber: "CLM-ABC12345"
                      claimantMemberNumber: "CM-1001"
                      claimType: MOTOR
                      status: ASSIGNED
                      description: Car accident on Main Street involving two vehicles
                      createdAt: "2024-01-15T10:30:00"
                      updatedAt: "2024-01-15T11:00:00"
                    message: Claim assigned successfully
                    timestamp: "2024-01-15T11:00:00.000Z"
        '400':
          description: Invalid state or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: Claim or staff member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/claims/{claimNumber}/assessments:
    post:
      tags:
        - Assessments
      summary: Create a new assessment
      description: Create an assessment for a claim to determine approval, rejection, or request for more information
      operationId: createAssessment
      parameters:
        - name: claimNumber
          in: path
          required: true
          schema:
            type: string
          description: Public claim number to assess
          example: CLM-ABC12345
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
            examples:
              approved:
                value:
                  claimNumber: CLM-ABC12345
                  staffNumber: STAFF-1001
                  assessmentType: INITIAL_REVIEW
                  description: Initial assessment completed. Claim appears valid and within policy terms.
                  details: All documentation provided. No red flags identified.
                  estimatedAmount: 5000.00
                  result: APPROVED
              moreInfo:
                value:
                  claimNumber: CLM-ABC12345
                  staffNumber: STAFF-1001
                  assessmentType: INITIAL_REVIEW
                  description: Additional information needed for approval.
                  details: Police report missing. Request copy from claimant.
                  estimatedAmount: 5000.00
                  result: MORE_INFO_REQUIRED
      responses:
        '201':
          description: Assessment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssessment'
              examples:
                default:
                  value:
                    success: true
                    data:
                      id: 66666666-6666-6666-6666-666666666666
                      claimNumber: "CLM-ABC12345"
                      staffNumber: STAFF-1001
                      assessmentType: INITIAL_REVIEW
                      description: Initial assessment completed
                      details: All documentation provided
                      estimatedAmount: 5000.00
                      settledAmount: null
                      result: APPROVED
                      createdAt: "2024-01-15T11:30:00"
                      updatedAt: "2024-01-15T11:30:00"
                    message: Assessment created successfully
                    timestamp: "2024-01-15T11:30:00.000Z"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: Claim or staff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
    get:
      tags:
        - Assessments
      summary: Get assessments for a claim
      description: Retrieve all assessments associated with a specific claim
      operationId: getAssessmentsByClaim
      parameters:
        - name: claimNumber
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Public claim number to retrieve assessments for
          example: CLM-ABC12345
      responses:
        '200':
          description: Assessments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssessmentList'
              examples:
                default:
                  value:
                    success: true
                    data:
                      - id: 66666666-6666-6666-6666-666666666666
                        claimNumber: "CLM-ABC12345"
                        staffNumber: STAFF-1001
                        assessmentType: INITIAL_REVIEW
                        description: Initial assessment completed
                        details: All documentation provided
                        estimatedAmount: 5000.00
                        settledAmount: null
                        result: APPROVED
                        createdAt: "2024-01-15T11:30:00"
                        updatedAt: "2024-01-15T11:30:00"
                    message: OK
                    timestamp: "2024-01-15T11:30:00.000Z"
        '404':
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
  /api/v1/assessments/{assessmentId}:
    get:
      tags:
        - Assessments
      summary: Get assessment by ID
      description: Retrieve a specific assessment's details
      operationId: getAssessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Assessment ID
          example: 66666666-6666-6666-6666-666666666666
      responses:
        '200':
          description: Assessment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssessment'
        '404':
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/staff/{staffNumber}/assessments:
    get:
      tags:
        - Assessments
      summary: Get assessments by staff member
      description: Retrieve all assessments performed by a specific staff member
      operationId: getAssessmentsByStaff
      parameters:
        - name: staffNumber
          in: path
          required: true
          schema:
            type: string
          description: Public staff number
          example: STAFF-1001
      responses:
        '200':
          description: Assessments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssessmentList'
        '404':
          description: Staff member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/assessments/{assessmentId}/settlement:
    patch:
      tags:
        - Assessments
      summary: Update settled amount
      description: Update the settled amount for an assessment
      operationId: updateSettledAmount
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Assessment ID
          example: 66666666-6666-6666-6666-666666666666
        - name: settledAmount
          in: query
          required: true
          schema:
            type: number
            format: double
          description: The settled amount for the assessment
          example: 4800.00
      responses:
        '200':
          description: Settled amount updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssessment'
              examples:
                default:
                  value:
                    success: true
                    data:
                      id: 66666666-6666-6666-6666-666666666666
                      claimNumber: "CLM-ABC12345"
                      staffNumber: STAFF-1001
                      assessmentType: INITIAL_REVIEW
                      description: Initial assessment completed
                      details: All documentation provided
                      estimatedAmount: 5000.00
                      settledAmount: 4800.00
                      result: APPROVED
                      createdAt: "2024-01-15T11:30:00"
                      updatedAt: "2024-01-15T12:15:00"
                    message: Settled amount updated successfully
                    timestamp: "2024-01-15T12:15:00.000Z"
        '404':
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
components:
  schemas:
...
    CreateClaimRequest:
      type: object
      required:
        - claimantMemberNumber
        - claimType
        - description
      properties:
        claimantMemberNumber:
          type: string
          description: Claimant member number used to identify the claimant.
          example: CM-1001
        claimType:
          $ref: '#/components/schemas/ClaimType'
        description:
          type: string
          minLength: 1
          description: Detailed description of the claim
          example: "Car accident on Main Street involving two vehicles. Police report filed."
      description: Request payload for creating a new claim
    ClaimResponse:
      type: object
      properties:
        claimNumber:
          type: string
          description: Unique claim number generated by the system
          example: "CLM-ABC12345"
        claimantMemberNumber:
          type: string
          description: Claimant member number used for customer-facing tracking
          example: CM-1001
        claimType:
          $ref: '#/components/schemas/ClaimType'
        status:
          $ref: '#/components/schemas/ClaimStatus'
        description:
          type: string
          description: Description of the claim
          example: "Car accident on Main Street involving two vehicles"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the claim was created
          example: "2024-01-15T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the claim was last updated
          example: "2024-01-15T10:30:00"
      description: Response payload containing claim details
...
    ClaimType:
      type: string
      enum:
        - MOTOR
        - PROPERTY
      description: |-
        Type of insurance claim
        - MOTOR: Motor vehicle insurance claim
        - PROPERTY: Property damage insurance claim
      example: MOTOR
    ClaimStatus:
      type: string
      enum:
        - SUBMITTED
        - ASSIGNED
        - ASSESSMENT_IN_PROGRESS
        - MORE_INFO_REQUESTED
        - APPROVED
        - REJECTED
        - SETTLEMENT_IN_PROGRESS
        - CLOSED
      description: |-
        Current status of the claim in its lifecycle
        - SUBMITTED: Initial claim submission
        - ASSIGNED: Claim assigned to staff member
        - ASSESSMENT_IN_PROGRESS: Under assessment
        - MORE_INFO_REQUESTED: Awaiting additional information
        - APPROVED: Claim approved
        - REJECTED: Claim rejected
        - SETTLEMENT_IN_PROGRESS: Settlement being processed
        - CLOSED: Claim closed
      example: SUBMITTED
...
    ApiResponseClaimResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        data:
          $ref: '#/components/schemas/ClaimResponse'
        message:
          type: string
          description: Response message
          example: "Claim created successfully"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2024-01-15T10:30:00.000Z"
      description: API response wrapper for ClaimResponse data
...
    ApiResponseClaimResponseList:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ClaimResponse'
          description: List of claims
        message:
          type: string
          description: Response message
          example: "OK"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2024-01-15T10:30:00.000Z"
      description: API response wrapper for list of ClaimResponse data
...
    CreateAssessmentRequest:
      type: object
      required:
        - claimNumber
        - staffNumber
        - assessmentType
        - description
        - estimatedAmount
        - result
      properties:
        claimNumber:
          type: string
          description: Public claim number being assessed
          example: CLM-ABC12345
        staffNumber:
          type: string
          description: Public staff number of the assessor
          example: STAFF-1001
        assessmentType:
          type: string
          minLength: 1
          description: Type of assessment being performed
          example: INITIAL_REVIEW
        description:
          type: string
          minLength: 1
          description: Description of the assessment findings
          example: "Initial assessment completed. Claim appears valid."
        details:
          type: string
          description: Additional detailed notes on the assessment
          example: "All documentation provided. No red flags identified."
        estimatedAmount:
          type: number
          format: double
          description: Estimated settlement amount
          example: 5000.00
        result:
          $ref: '#/components/schemas/AssessmentResult'
      description: Request payload for creating a new assessment
    AssessmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the assessment
          example: 66666666-6666-6666-6666-666666666666
        claimNumber:
          type: string
          description: Claim number associated with this assessment
          example: "CLM-ABC12345"
        staffNumber:
          type: string
          description: Public staff number of the assessor
          example: STAFF-1001
        assessmentType:
          type: string
          description: Type of assessment
          example: INITIAL_REVIEW
        description:
          type: string
          description: Description of the assessment findings
          example: "Initial assessment completed"
        details:
          type: string
          description: Detailed notes on the assessment
          example: "All documentation provided"
        estimatedAmount:
          type: number
          format: double
          description: Estimated settlement amount
          example: 5000.00
        settledAmount:
          type: number
          format: double
          nullable: true
          description: Actual settled amount
          example: 4800.00
        result:
          $ref: '#/components/schemas/AssessmentResult'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the assessment was created
          example: "2024-01-15T11:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the assessment was last updated
          example: "2024-01-15T11:30:00"
      description: Response payload containing assessment details
    AssessmentResult:
      type: string
      enum:
        - APPROVED
        - REJECTED
        - MORE_INFO_REQUIRED
      description: |-
        Result of the assessment
        - APPROVED: Claim approved for settlement
        - REJECTED: Claim rejected
        - MORE_INFO_REQUIRED: Additional information needed from claimant
      example: APPROVED
    ApiResponseAssessment:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        data:
          $ref: '#/components/schemas/AssessmentResponse'
        message:
          type: string
          description: Response message
          example: "Assessment created successfully"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2024-01-15T11:30:00.000Z"
      description: API response wrapper for Assessment data
    ApiResponseAssessmentList:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentResponse'
          description: List of assessments
        message:
          type: string
          description: Response message
          example: "OK"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2024-01-15T11:30:00.000Z"
      description: API response wrapper for list of Assessment data
    ApiResponseError:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: false
        data:
          type: 'null'
          description: No data in error response
        message:
          type: string
          description: Error message
          example: "Resource not found"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2024-01-15T10:30:00.000Z"
      description: API response wrapper for error responses
...